requirements.txt
fastapi
uvicorn
matplotlib
numpy
pillow
